package config2 single index file, import dist folder
