+++ packages/fxhash-config2/package.json
@@ -3,13 +3,15 @@
   "version": "1.0.0",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "files": ["dist/**/*"],
   "license": "MIT",
   "scripts": {
     "clean": "rm -rf dist",
-    "build": "npm run clean && tsc --declaration && tsc-alias",
-    "build:watch": "npm run clean && tsc && (concurrently \"tsc -w --declaration\" \"tsc-alias -w\")",
+    "build": "npm run clean && tsc --declaration",
+    "build:watch": "tsc && (concurrently \"tsc -w --declaration\" \"tsc-alias -w\")",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
+    "typescript": "^5.1.6"
   }
 }
