+++ packages/fxhash-config2/tsconfig.json
@@ -18,7 +18,7 @@
     "noUnusedParameters": true,
     "noImplicitAny": false,
     "noImplicitThis": false,
-    "strictNullChecks": false,
+    "strictNullChecks": false
   },
   "include": ["src/**/*", "__tests__/**/*"],
   "exclude": ["src/static/**"]
